Bookman 1.0.1
Fixed bug whereby changing inspection elements or manufacturers via the Preferences window would result in crashes if work instructions or rules existed which relied on the changed entities.

Implement a right-click menu item in the rules sheet to change the work instruction associated with a rule.

Inspection element numbers are now 5 digits wide.

Stop protecting the Cables inspection element and create a separate special inspection element called Xxx with inspection element number 99999 which is protected.

Allow sorting of groups tables and rules tables by clicking on the various JTable column headers.

Fix a crash which resulted when hiding a rule in a rule sheet.

Change rule sheet management to allow a rule sheet to "tolerate" having a rule with an inspection element or manufacturer which is incompatible with its defined IE or manufacturer. Note that rules which are in the "wrong" sheet get moved to the "right" sheet when restoring from a save file. Consequently, this "tolerance" is quite limited - i.e. the copied rules should have their inspection elements and/or manufacturers changed to be compatible with the sheet that they are now in by using the new right-click menu item to change their work instructions.