Implement a FormattingLinkedList class which derives from the Java LinkedList class and provides a toString() method which returns a formatted 'display' of the contents of the list. Other than constructors, there are no other public or protected methods in this class.
Augment the new ValidatedJTextField class to pass a parameter to its storeNewValue method which allows the implementation to distinguish between stores which result from the user pressing return (i.e. detected via an invocation of the ActionListener.actionPerformed method) or other causes (currently only loss of focus in the text field).