Bookman 1.1.3
Replace de-select code in the rules spreadsheet with code that remembers the REETs which have been selected and then just de-selects them explicitly when the time comes to do a de-selection.

Rework the inspection element number and name parsing code to avoid booms when validating user input when they change the number or name of an inspection element.

Reduce the minimum size of most of the various Bookman windows.  Adjust the preferred size on a few to make sure that look reasonable when first opened.

Rework how rules are duplicated to avoid a boom (no idea how this every worked).

Add a validation method to the RSC class which checks that the rules, rule clauses and REETs in the rules in the sheet contain no duplicates (would have caught the duplication bug described above).