Bookman 1.6.16
Implement the ability to change presentation category code letters. Changing the code letter for a category implicitly changes the code letter at the start of any presentation weight which was using the former code letter. For example, if a presentation category has its code letter changed from 'K' to 'X' then any presentation weights in inspection elements or groups which used to start with a 'K' will now start with an 'X'.

Note that when changing a category's code letter, the new code letter must not already be in use. The effect of this rule is that this mechanism cannot be used to merge two presentation categories. For example, if there are two categories with code letters 'A' and 'B' then Bookman will not allow the first category's code letter to be changed from 'A' to 'B' or the second category's code letter to be changed from 'B' to 'A'. Not having this rule just seemed to make it to easy to accidentally merge two categories - an action which could be a fair bit of work to reverse.

Various cleanup.