Gowing cleanup including simplifying to extract of strings from bundles.
New Ix class that can be used to turn collections of things into collections of things wrapped in an Ix instance where the wrapper knows its index in the original collection. This tends to avoid if not eliminate the practice of either spinning through a collection using an integer index style for loop or spinning through a collection using the more typical "for ( Abc item : items )" approach while (trying) to keep a separate indexing variable in sync so that elements can later be found by their index.