Added a lot of final, @NotNull, and @Nullable attributes.
Basic sanity on selecting data category and repository connections in the editor window (this took way too long to settle down - sigh).

The usual mountain of cleanup.

There are a bunch of classes that represent abandoned ideas that arose during the process of figuring out how to do selection and management of source attributes (the data category, repository connector, data flavour, etc stuff in the editor window). These will disappear soon (probably in the next commit).