Add fairly basic "regular verbose" and "super verbose" mechanism to make it a bit easier to figure out why the "finish unpacking" phase of a Gowing unpack operation is failing.
Add an unchecked GowingUnpackingDeadlockedException which is thrown when the "finish unpacking" phase of a Gowing unpack operation reaches a point where a round of finishing doesn't reduce the number of unfinished unpackables. This is usually because two packables are waiting for each other to finish (generally if not always because one or both of the packables have a "cannot finish until the other is finished" dependency that doesn't really need to be there, or the finishing strategy being used is flawed - roughly speaking, if it was possible to build the data structure one step-at-a-time in the first place then there must be a way to unpack it while finishing it one-step-at-a-time).

Cleanup of various aspects of the Gowing facility and elsewhere.