Stop using various APIs that were deprecated in Java 9 (mostly AWT APIs).
Implement the FlexiGridDivider class which provides a way to include "divider lines" in components being managed by the FlexiGridLayoutManager (strictly speaking, this is intended for use with the FlexiGridPanelModel although the actual support for it is in the layout manager part of the flexi-grid facility).

Various cleanup including a bit of reorganizing of classes and methods.