Fairly major rework to provide clear visual feedback when tabbing between selectable components in the editor pane.
Rework to get included field handling behave more consistently (when a field is included, the data field and, if it exists, the copy-in-value stuff is enabled; when the field is not included then the data field and, if it exists, the copy-in-value stuff is disabled).

Fix ReferenceCard so that when a particular census instance is selected, only the census data fields associated with that census instance appear in the editor pane.