Eliminate the distinction between baptism, marriage and burial cards and combine them all into a single ReferenceCard.
Work towards a reference referring to WikiTreeIds rather than people.

Create a WikiTreeDbWtid class to manage fields whose value is a WikiTreeId. The difference between this class and the WikiTreeDbPerson class is that the value that the user edits in this class is the WikiTreeId whereas the value that the user edits in the WikiTreeDbPerson class is how the person is named.