Create a reasonably general purpose listener facility (ObtuseListenerProxy and SimpleObtuseListenerManager) which is intended to make it easy to add listener functionality to classes that are possibly best monitored via events.
Use the listener facility to add selection notifications to the wrapped combo-box and button-model classes.

Implement the WrappedLabelPanel class that is essentially a JPanel that can wrap either a JLabel or a  WrappedComboBoxSelectorPanel. This makes it easier to support the degenerate case where a conceptual combo-box actually has only one value.