Cleanup and attempts to increase the speed of Gowing (especially on the unPacking side of things). It turns out that Gowing itself is actually fairly fast. The bottleneck seems to be in the user classes which sometimes have to do a bit of a dance to get things done. I'm looking at some ideas like a better approach to handling collections. Something for the future.
Implemented TypedGowingUnPacker which is a somewhat simplified API. I'm still not happy with the amount of boilerplate code that needs to be written in userland. Another work in progress.

MonikerOwner is just a simple interface that marks classes which have a {@code String getMoniker()} method. The intent is to provide a way to clearly mark classes which can provide a possibly less than formal but still useful name for each instance. Not sure how useful it is but it's part of the Pipestone family now.

Minor change to the Logger class to make it possible to get the name of the file where a particular Logger instance is archiving what it writes. See getMirrorName() for (not much) more info.

Lots of minor cleanup in a vain attempt to reduce the noise that comes out of IntelliJ's automated code analysis.