Considerably reworked AutoSizingValidatedJTextField class (required to get it to work in more than just the simplest of cases).
Considerably reworked FlexiGridCache1 class (also required to get it to work in more than just the simplest of cases). This version has the notion of a filler column which is where excess whitespace is placed. This still needs more thought (not to mention that if there are filler columns then there should probably be filler rows).

Various relatively minor cleanup.