Create the ObtuseVideoUtils class (hopefully, the start of something big).
Implement the AsyncStreamCaptor class which asynchronously captures the contents of an {@link InputStream}.

Implement a makeTextImageIcon method in the ImageButton class that creates an {@link ImageIcon} containing a single row of text.

Implement two more readEntireFile methods in the ObtuseUtil class that will read the contents of an arbitrarily large file (i.e. the caller does not need to provide an upper bounds on the file system). This is probably a bad idea considering that it does not provide any way to avoid an out-of-memory condition if the file is absurdly large. Already marked as deprecated.

Implement the UnixSignals class which provides various useful constants for dealing with Unix signals. Also, provides methods that make it easy to tear apart the 'exit status' of a Unix process.