Add a new date format in DateUtils (MMM_DD_HHMMSS_YYYY) to handle dates like .Mar 15, 12:15:00 2012".
Move many of the "describe this GUI thing" from LinearLayoutUtil to the new ObtuseGuiUtils utility class.

Implement a simple interface to allow the measurements produced by the Measure class to be sorted in various ways.

Implement the QuickLogger class which provides a much simpler logging facility that the existing Logger class.

Implement Page-Up and Page-Down key support in the VirtualScrollablePanel facility. It's still a bit clunky but it's a lot better than having to scroll or mouse one's way through long scrollable panels.

Plenty of minor cleanup.