Implemented a file selection dialog which is intended to be more customizable and less confusing to use than either JFileChooser or FileDialog.
Implement the DiagonalLayout manager which is intended as a test bed for future layout managers and a demonstration of a very basic layout manager.

Implement the ObtuseFlowLayoutManager which is intended as a substitute for the Java FlowLayout manager. It is hopefully more stable and predictable than FlowLayout.

Add more filesystem-related 16x16 icons (used by the new ObtuseFileSelector class).

Implement the VspSelectionUtils class. Not sure if it has an actual purpose in the long term.

Implement the BasicVirtualScrollablePanelSelectionManager class which is intended to be used with the VirtualScrollablePanel facility.

Minor cleanup all over the place.