Relatively minor cleanup (often removing generics-related warnings) of DateSortController, FieldValueMap. IndexFieldHelper, LancotIndexFileManager, LancotIndexManager, MultiFieldValueMap, SingleFieldValueMap, LancotLibraryRoot, LancotMediaItem, LancotMediaItemPanel, LancotFiles, LancotGuiAction, LancotMediaProxy, LancotScrollableImagesPanelModel, LancotScrollableImagesRowView, LancotTagTreeScrollablePanelModel, and MediaBrowserWindow.
Introduce the notion of protected tags to the LancotTags classes.

Add some test images to xyz.kenosee.lancot.resources and move the folder image variants to the Pipestone project.

Set a dock image in Lancot class.

Rework the ImageAppearanceControlPanel to make it easier to use in other places.

Support the adding, removing and renaming of tags in LancotTagElementModel and LancotTagElementView

Implement an expression parser and evaluator. This is intended to be used by a future search interface. The actual expression parser and evaluator should and probably will be cleaned up and moved to the Pipestone project.

Change DEFAULT_THREAD_COUNT from 1 to 8 in AsyncThumbnailGenerator.

Rework the exporting of media items in LancotMediaBrowserPanel and LancotMediaItemsExportManager.

Implement GetSingleValueDialog which is a simple popup window for prompting for single values (probably not ready for prime time; will probably be moved to the Pipestone class once it is ready for prime time).